EQS
added behavior for the bot to go look for food when stamina is low
