--- Source/AI_Example/Private/AIE_Asset_Paths.cpp
@@ -1,10 +1,0 @@
-// Copyright Jordan Duncan 2016
-
-#include "AI_Example.h"
-#include "AIE_Asset_Paths.h"
-
-// bot items
-const FString FAIE_Asset_Paths::DefaultBotMesh = FString("SkeletalMesh'/Game/Mannequin/Character/Mesh/SK_Mannequin.SK_Mannequin'");
-const FString FAIE_Asset_Paths::DefaultBotAnim = FString("AnimBlueprint'/Game/Mannequin/Animations/ThirdPerson_AnimBP.ThirdPerson_AnimBP_C'");
-// Food items
-const FString FAIE_Asset_Paths::DefaultFoodMesh = FString("StaticMesh'/Game/InfinityBladeGrassLands/Environments/Misc/Exo_Deco01/StaticMesh/SM_WoodenBarrel_Intact.SM_WoodenBarrel_Intact'");
