+++ Source/AI_Example/Public/AI/BTNodes/BTTasks/AIE_GoToRandom_BTTaskNode.h
@@ -1,0 +1,26 @@
+// Copyright Jordan Duncan 2016
+
+#pragma once
+
+#include "AIE_BTTaskNode.h"
+#include "AIE_AIController.h"
+#include "AIE_GoToRandom_BTTaskNode.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class AI_EXAMPLE_API UAIE_GoToRandom_BTTaskNode : public UAIE_BTTaskNode
+{
+	GENERATED_BODY()
+	// constructor
+	UAIE_GoToRandom_BTTaskNode(const FObjectInitializer& ObjectInitializer);
+public:
+	// execution
+	virtual EBTNodeResult::Type ExecuteTask(UBehaviorTreeComponent& OwnerComp, uint8* NodeMemory) override;
+	// bool used to force return success of task execution
+	bool bForceSuccess = true;
+	// used to determin the search range for our new location
+	float searchRadius = 3000.0f;
+
+};
