+++ Source/AI_Example/Public/Strings/AIE_Asset_Paths.h
@@ -1,0 +1,18 @@
+// Copyright Jordan Duncan 2016
+
+#pragma once
+#include "AIE_Asset_Paths.generated.h"
+/**
+ * 
+ */
+USTRUCT()
+struct AI_EXAMPLE_API FAIE_Asset_Paths
+{
+	GENERATED_USTRUCT_BODY()
+public:
+	// bot assets
+	static const FString DefaultBotMesh;
+	static const FString DefaultBotAnim;
+	// Items
+	static const FString DefaultFoodMesh;
+};
